requirements.txt
# Weather ETL Pipeline Requirements

# HTTP requests
requests>=2.31.0

# Scheduling
APScheduler>=3.10.4

# Environment variables
python-dotenv>=1.0.0

# Type hints (for Python < 3.9)
typing-extensions>=4.8.0
